Solving simple-post 	* Solving single post view - not complete yet
