+++ .eslintrc
@@ -5,6 +5,15 @@
   },
   "rules": {
     "import/prefer-default-export": 0,
-    "no-restricted-syntax": 0
+    "no-restricted-syntax": 0,
+    "react/jsx-filename-extension": [
+      1,
+      {
+          "extensions": [
+              ".js",
+              ".jsx"
+          ]
+      }
+    ]
 }
 }
