+++ simple-post/frontend/package.json
@@ -13,6 +13,7 @@
     "prop-types": "^15.6.1",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
+    "react-router-dom": "^5.0.1",
     "react-scripts": "1.1.4"
   },
   "devDependencies": {
