+++ simple-post/frontend/src/components/Post.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 
 const Post = ({ post }) => {
   return (
     <div key={post.id}>
       <h4>{post.id}</h4>
-      <div>Title: {post.title}</div>
+      <div>Title: <Link to={`/post/${post.id}`}>{post.title}</Link></div>
     </div>
   );
 };
