+++ simple-post/frontend/src/lib/API.js
@@ -1,6 +1,6 @@
-const get = async (url) => {
+const get = async (path) => {
   // eslint-disable-next-line no-undef
-  const response = await fetch(url, {
+  const response = await fetch(`http://localhost:3001${path}`, {
     method: 'GET', // *GET, POST, PUT, DELETE, etc.
     mode: 'cors', // no-cors, cors, *same-origin
     cache: 'no-cache', // *default, no-cache, reload, force-cache, only-if-cached
